refactor multi server search to seperate file
